Extract out rigidBodyManipulatorMexConversions.h. Use the fromMex for KinematicsCache in collision mex functions so that runtime type checking is performed (i.e. they no longer blindly assume that the DrakePointer they get is a DrakePointer to KinematicsCache<double>).
