+++ drake/systems/plants/rigidBodyManipulatorMexFunctions.h
@@ -29,6 +29,5 @@ DLLEXPORT void geometricJacobianDotTimesVmex(int nlhs, mxArray *plhs[], int nrhs
 DLLEXPORT void geometricJacobianmex(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[]);
 DLLEXPORT void massMatrixmex(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[]);
 DLLEXPORT void dynamicsBiasTermmex(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[]);
-DLLEXPORT void contactConstraintsmex(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[]);
 
 #endif //DRAKE_DRAKEMEXFUNCTIONS_H
